implemented comments service and controller

+++ src/api/db/models/Movie.ts
@@ -30,7 +30,7 @@ Movie.init(
             allowNull: false,
         },
         rating: {
-            type: DataTypes.SMALLINT,
+            type: DataTypes.FLOAT,
             allowNull: false,
         },
         releaseDate: {
